refactor(agent): split durable sync policy modules
